Layout changes and Merge fixes
