+++ SPEED-app/SPEED-frontend/app/view-articles/page.tsx
@@ -1,14 +1,12 @@
 "use client"; // Declare this file as a Client Component
 
 import React, { useEffect, useState } from 'react';
-import PageTemplate from '../page-template/page';
 import axios from 'axios';
 
 const Page = () => {
   const [articles, setArticles] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-
   const apiUrl = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:8082';
 
   // Fetch approved articles when the component mounts
@@ -27,37 +25,56 @@ const Page = () => {
     fetchApprovedArticles();
   }, []);
 
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>{error}</p>;
+  // Handle loading and error states
+  if (loading) {
+    return (
+      <main className="p-6 flex-1 bg-gray-100">
+        <h2 className="text-2xl font-semibold mb-4">
+          View Articles
+        </h2>
+        <p className="mb-4">
+          Use SPEED to search and analyse claims about various Software Engineering practices.
+        </p>
+        <p className="mb-4 text-gray-400">Loading...</p>
+      </main>
+    );
+  }
 
+  if (error) {
+    return (
+      <main className="p-6 flex-1 bg-gray-100">
+        <h2 className="text-2xl font-semibold mb-4">
+          View Articles
+        </h2>
+        <p className="mb-4">
+          Use SPEED to search and analyse claims about various Software Engineering practices.
+        </p>
+        <p className="mb-4 text-red-500">An error occured while loading articles</p>
+      </main>
+    );
+  }
+
+  // Render the page content with articles
   return (
-    <PageTemplate
-      pageContent={
-        <main className="p-6 flex-1 bg-gray-100">
-          <h2 className="text-2xl font-semibold mb-4">View Articles</h2>
-          <p className="mb-4">
-            Use SPEED to search and analyse claims about various Software Engineering practices.
-          </p>
-          {/* Map through approved articles and render each one */}
-          {articles.length === 0 ? (
-            <p>No articles to display.</p>
-          ) : (
-            articles.map((article: any) => (
-              <div key={article._id} className="bg-white p-4 rounded shadow mb-4">
-                <h3 className="text-xl font-semibold mb-2">{article.title}</h3>
-                <p className="mb-2">
-                  <strong>Author:</strong> {article.author || "Unknown"}
-                </p>
-                <p>{article.description || "No description available"}</p>
-                <p>{article.published_date || "No publication date"}</p>
-                <p>{article.publisher || "Unknown publisher"}</p>
-                <p>{article.updated_date || "No update date"}</p>
-              </div>
-            ))
-          )}
-        </main>
-      }
-    />
+    <main className = "p-6 flex-1 bg-gray-100">
+      <h2 className="text-2xl font-semibold mb-4">
+          View Articles
+      </h2>
+      <p className="mb-4">
+          Use SPEED to search and analyse claims about various Software Engineering practices.
+      </p>
+      {/* Map through articles and render each one */}
+      {articles.map((article:any) => (
+          <div key={article.isbn} className="bg-white p-4 rounded shadow mb-4">
+          <h3 className="text-xl font-semibold mb-2">{article.title}</h3>
+          <p className="mb-2"><strong>Author:</strong> {article.author}</p>
+          <p>{article.description}</p>
+          <p>{article.published_date}</p>
+          <p>{article.publisher}</p>
+          <p>{article.update_date}</p>
+          </div>
+        ))}
+    </main>
   );
 };
